+++ src/main/java/uk/co/ogauthority/pathfinder/controller/project/collaborationopportunites/forwardworkplan/ForwardWorkPlanCollaborationOpportunityController.java
@@ -22,11 +22,13 @@ import uk.co.ogauthority.pathfinder.config.file.FileDeleteResult;
 import uk.co.ogauthority.pathfinder.config.file.FileUploadResult;
 import uk.co.ogauthority.pathfinder.controller.file.FileDownloadService;
 import uk.co.ogauthority.pathfinder.controller.file.PathfinderFileUploadController;
+import uk.co.ogauthority.pathfinder.controller.project.annotation.AllowProjectContributorAccess;
 import uk.co.ogauthority.pathfinder.controller.project.annotation.ProjectFormPagePermissionCheck;
 import uk.co.ogauthority.pathfinder.controller.project.annotation.ProjectStatusCheck;
 import uk.co.ogauthority.pathfinder.controller.project.annotation.ProjectTypeCheck;
 import uk.co.ogauthority.pathfinder.exception.AccessDeniedException;
 import uk.co.ogauthority.pathfinder.model.entity.project.ProjectDetail;
+import uk.co.ogauthority.pathfinder.model.entity.project.collaborationopportunities.forwardworkplan.ForwardWorkPlanCollaborationOpportunity;
 import uk.co.ogauthority.pathfinder.model.enums.ValidationType;
 import uk.co.ogauthority.pathfinder.model.enums.audit.AuditEvent;
 import uk.co.ogauthority.pathfinder.model.enums.project.ProjectStatus;
@@ -38,6 +40,8 @@ import uk.co.ogauthority.pathfinder.mvc.ReverseRouter;
 import uk.co.ogauthority.pathfinder.service.audit.AuditService;
 import uk.co.ogauthority.pathfinder.service.controller.ControllerHelperService;
 import uk.co.ogauthority.pathfinder.service.file.ProjectDetailFileService;
+import uk.co.ogauthority.pathfinder.service.project.OrganisationGroupIdWrapper;
+import uk.co.ogauthority.pathfinder.service.project.ProjectSectionItemOwnershipService;
 import uk.co.ogauthority.pathfinder.service.project.collaborationopportunities.forwardworkplan.ForwardWorkPlanCollaborationCompletionService;
 import uk.co.ogauthority.pathfinder.service.project.collaborationopportunities.forwardworkplan.ForwardWorkPlanCollaborationOpportunitiesSummaryService;
 import uk.co.ogauthority.pathfinder.service.project.collaborationopportunities.forwardworkplan.ForwardWorkPlanCollaborationOpportunityFileLinkService;
@@ -52,6 +56,7 @@ import uk.co.ogauthority.pathfinder.util.validation.ValidationResult;
 @Controller
 @ProjectStatusCheck(status = ProjectStatus.DRAFT)
 @ProjectFormPagePermissionCheck
+@AllowProjectContributorAccess
 @ProjectTypeCheck(types = ProjectType.FORWARD_WORK_PLAN)
 @RequestMapping("/project/{projectId}/forward-work-plan/collaboration-opportunities")
 public class ForwardWorkPlanCollaborationOpportunityController extends PathfinderFileUploadController {
@@ -64,6 +69,7 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
   private final ForwardWorkPlanCollaborationOpportunitiesSummaryService forwardWorkPlanCollaborationOpportunitiesSummaryService;
   private final ForwardWorkPlanCollaborationCompletionService forwardWorkPlanCollaborationCompletionService;
   private final ControllerHelperService controllerHelperService;
+  private final ProjectSectionItemOwnershipService projectSectionItemOwnershipService;
 
   @Autowired
   public ForwardWorkPlanCollaborationOpportunityController(
@@ -76,8 +82,8 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
       ForwardWorkPlanCollaborationCompletionService forwardWorkPlanCollaborationCompletionService,
       ProjectDetailFileService projectDetailFileService,
       ControllerHelperService controllerHelperService,
-      FileDownloadService fileDownloadService
-  ) {
+      FileDownloadService fileDownloadService,
+      ProjectSectionItemOwnershipService projectSectionItemOwnershipService) {
     super(projectDetailFileService, fileDownloadService);
     this.forwardWorkPlanCollaborationOpportunityModelService = forwardWorkPlanCollaborationOpportunityModelService;
     this.forwardWorkPlanCollaborationOpportunityFileLinkService = forwardWorkPlanCollaborationOpportunityFileLinkService;
@@ -87,6 +93,7 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
     this.forwardWorkPlanCollaborationOpportunitiesSummaryService = forwardWorkPlanCollaborationOpportunitiesSummaryService;
     this.forwardWorkPlanCollaborationCompletionService = forwardWorkPlanCollaborationCompletionService;
     this.controllerHelperService = controllerHelperService;
+    this.projectSectionItemOwnershipService = projectSectionItemOwnershipService;
   }
 
   @GetMapping("/setup")
@@ -221,6 +228,7 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
                                                    @PathVariable("opportunityId") Integer opportunityId,
                                                    ProjectContext projectContext) {
     var opportunity = forwardWorkPlanCollaborationOpportunityService.getOrError(opportunityId);
+    checkIfUserHasAccessToCollaborationOpportunity(opportunity);
     return getCollaborationOpportunityModelAndView(
         (ForwardWorkPlanCollaborationOpportunityForm) forwardWorkPlanCollaborationOpportunityService.getForm(opportunity),
         projectContext.getProjectDetails()
@@ -235,6 +243,7 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
                                                      ValidationType validationType,
                                                      ProjectContext projectContext) {
     var opportunity = forwardWorkPlanCollaborationOpportunityService.getOrError(opportunityId);
+    checkIfUserHasAccessToCollaborationOpportunity(opportunity);
     bindingResult = forwardWorkPlanCollaborationOpportunityService.validate(form, bindingResult, validationType);
 
     return controllerHelperService.checkErrorsAndRedirect(
@@ -267,9 +276,11 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
                                                             @PathVariable("opportunityId") Integer opportunityId,
                                                             @PathVariable("displayOrder") Integer displayOrder,
                                                             ProjectContext projectContext) {
+    var opportunity = forwardWorkPlanCollaborationOpportunityService.getOrError(opportunityId);
+    checkIfUserHasAccessToCollaborationOpportunity(opportunity);
     return forwardWorkPlanCollaborationOpportunityModelService.getRemoveCollaborationOpportunityConfirmationModelAndView(
         projectId,
-        opportunityId,
+        opportunity,
         displayOrder
     );
   }
@@ -280,6 +291,7 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
                                                      @PathVariable("displayOrder") Integer displayOrder,
                                                      ProjectContext projectContext) {
     final var opportunity = forwardWorkPlanCollaborationOpportunityService.getOrError(opportunityId);
+    checkIfUserHasAccessToCollaborationOpportunity(opportunity);
     forwardWorkPlanCollaborationOpportunityService.delete(opportunity);
 
     AuditService.audit(
@@ -369,4 +381,16 @@ public class ForwardWorkPlanCollaborationOpportunityController extends Pathfinde
     );
   }
 
+  private void checkIfUserHasAccessToCollaborationOpportunity(ForwardWorkPlanCollaborationOpportunity opportunity) {
+    if (!projectSectionItemOwnershipService.canCurrentUserAccessProjectSectionInfo(
+        opportunity.getProjectDetail(),
+        new OrganisationGroupIdWrapper(opportunity.getAddedByOrganisationGroup())
+    )) {
+      throw new AccessDeniedException(
+          String.format(
+              "User does not have access to the ForwardWorkPlanCollaborationOpportunity with id: %d",
+              opportunity.getId())
+      );
+    }
+  }
 }
